requirements.txt
librosa==0.9.2
pypinyin==0.46.0
playsound==1.3.0
#注意：如果直接用pip install playsound安装，Windows系统下会报错
#需要去原作者Github页面下载最新的playsound.py
#https://github.com/TaylorSMarks/playsound/blob/master/playsound.py
Pillow==9.0.1